Add README with instructions
Also updated Makefile to match the changed name of CA key generation.
Added a script to start all clients at once.
Made server be able to run "forever" to that it does not have to closed
for new shares to be sent, though 3 shares must be sent for the program
to aggregate them.